packages/modal-ocr/pyproject.toml
[project]
name = "scanfactory-modal-ocr"
version = "1.1.0"
description = "ScanFactory OCR service on Modal with PaddleOCR"
requires-python = ">=3.10"
dependencies = [
    "modal>=0.64.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
